Added style for return buttons, database, bootstrap files

+++ vue/vueAllMembers.php
@@ -13,49 +13,56 @@
     <?php if (isset($allMembers)) : ?>
 
 
-        <table>
+        <table class="all_members">
+
             <caption>Informations internaute</caption>
-            <tr>
-                <th scope="col">Libellé</th>
-                <th scope="col"></th>
-                <th scope="col">Actions</th>
-            </tr>
             <?php foreach ($allMembers as $key => $value) : ?>
+                <tr>
+                    <th scope="col">Libellé</th>
+                    <th scope="col"></th>
+                    <th scope="col">Actions</th>
+                </tr>
+
 
                 <tr>
                     <th scope="row">Login
-                        <button class="hidden-btn">Modifier</button>
+                        <a href="?action=allMembers&amp;memberId=<?= $value->getId_user() ?>" class="hidden-btn">Modifier
+                            le compte</a>
                     </th>
                     <td><?= $value->getLogin() ?></td>
-                    <td><a href="?action=allMembers&amp;memberId=<?= $value->getId_user() ?>">Modifier</a></td>
+                    <td><a href="?action=allMembers&amp;memberId=<?= $value->getId_user() ?>">Modifier le compte</a>
+                    </td>
+
                 </tr>
+
                 <tr>
                     <th scope="row">Email
-                        <button class="hidden-btn">Modifier</button>
+                        <a href="?action=allMembers&amp;deleteM=<?= $value->getId_user() ?>" onclick="return(confirm('Etes-vous sûr de vouloir supprimer?'));" class="hidden-btn">Supprimer
+                            le compte</a>
                     </th>
                     <td><?= $value->getEmail() ?></td>
-                    <td></td>
+                    <td><a href="?action=allMembers&amp;deleteM=<?= $value->getId_user() ?>" onclick="return(confirm('Etes-vous sûr de vouloir supprimer?'));">Supprimer le compte</a>
+
                 </tr>
                 <tr>
                     <th scope="row">Password
-                        <button class="hidden-btn">Modifier</button>
                     </th>
-                    <td><?= $value->getPassword() ?></td>
+                    <td><input type="password" value="<?= $value->getPassword() ?>"</td>
                     <td></td>
                 </tr>
                 <tr>
                     <th scope="row">Image
-                        <button class="hidden-btn">Modifier</button>
                     </th>
                     <td><a href="assets/img/uploads/<?= $value->getUser_image() ?>"><img
                                 src="assets/img/uploads/<?= $value->getUser_image() ?>" alt="Photo de profil"
                                 title="Cliquez pour agrandir" height="80em"/></a></td>
-                    <td><a href="?action=allMembers&amp;deleteM=<?= $value->getId_user() ?>">Supprimer le compte</a>
-                    </td>
+                    <td></td>
+
                 </tr>
 
 
             <?php endforeach; ?>
+
 
         </table>
 
@@ -99,6 +106,11 @@
 
     <?php endif; ?>
 <?php endif; ?>
+    <div class="back-page">
+        <div class="back-page"><a href="?action=home">Retour à l'accueil</a></div>
+
+
+    </div>
 <?php $template = ob_get_clean(); ?>
 
 <?php require 'templates/tempAccueil.php'; ?>
